seo-analyzer/requirements.txt
# SEO Analyzer - Dipendenze Python

# Core
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0
urllib3==2.2.0

# HTML Parsing & Analysis
html5lib==1.1
cssselect==1.2.0
pyquery==2.0.0

# Performance Analysis
selenium==4.18.1
playwright==1.42.0
lighthouse-python==0.1.0

# Image Analysis
Pillow==10.2.0
imagesize==1.4.1

# URL & Link Analysis
validators==0.22.0
tldextract==5.1.1

# SEO Specific
advertools==0.16.1
seo-position-tracking==0.1.0

# Data Processing
pandas==2.2.0
numpy==1.26.4

# Configuration & YAML
PyYAML==6.0.1
python-dotenv==1.0.1

# Reporting
jinja2==3.1.3
markdown==3.5.2
reportlab==4.1.0
weasyprint==61.2

# CLI & UI
click==8.1.7
rich==13.7.0
colorama==0.4.6
tqdm==4.66.2

# Async & Performance
aiohttp==3.9.3
asyncio==3.4.3

# Testing
pytest==8.0.2
pytest-cov==4.1.0
pytest-asyncio==0.23.5

# Utilities
python-dateutil==2.8.2
certifi==2024.2.2
charset-normalizer==3.3.2

